Uses data binding to eliminate findViewById
Data binding has the following benefits: 
Code is shorter, easier to read, and easier to maintain than code that uses findViewById(). 
Data and views are clearly separated. This benefit of data binding becomes increasingly important later in this course. 
The Android system only traverses the view hierarchy once to get each view, and it happens during app startup, not at runtime when the user is interacting with the app.
You get type safety for accessing views. (Type safety means that the compiler validates types while compiling, and it throws an error if you try to assign the wrong type to a variable.)

More @ Link: https://developer.android.com/codelabs/kotlin-android-training-data-binding-basics#2